Move otp helper to api module
